api/requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==21.4b2
bleach==4.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
cloudpickle==2.0.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
detectron2==0.6+cu102
entrypoints==0.4
executing==0.8.2
fonttools==4.29.1
future==0.18.2
fvcore==0.1.5.post20220119
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.43.0
hydra-core==1.1.1
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
iopath==0.1.9
ipykernel==6.8.0
ipython==8.0.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.2
oauthlib==3.2.0
omegaconf==2.1.1
opencv-python==4.5.5.62
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
portalocker==2.3.2
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
pydot==1.4.2
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
Send2Trash==1.8.0
six==1.16.0
stack-data==0.1.4
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
toml==0.10.2
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
yacs==0.1.8
zipp==3.7.0
